Add bin to encoding

Add bin parameters to an encoding

Usage

vl_bin_angle(
  spec,
  anchor = NULL,
  base = NULL,
  binned = NULL,
  divide = NULL,
  extent = NULL,
  maxbins = NULL,
  minstep = NULL,
  nice = NULL,
  step = NULL,
  steps = NULL,
  .object = NULL
)

vl_bin_color(
  spec,
  anchor = NULL,
  base = NULL,
  binned = NULL,
  divide = NULL,
  extent = NULL,
  maxbins = NULL,
  minstep = NULL,
  nice = NULL,
  step = NULL,
  steps = NULL,
  .object = NULL
)

vl_bin_description(
  spec,
  anchor = NULL,
  base = NULL,
  binned = NULL,
  divide = NULL,
  extent = NULL,
  maxbins = NULL,
  minstep = NULL,
  nice = NULL,
  step = NULL,
  steps = NULL,
  .object = NULL
)

vl_bin_detail(
  spec,
  anchor = NULL,
  base = NULL,
  binned = NULL,
  divide = NULL,
  extent = NULL,
  maxbins = NULL,
  minstep = NULL,
  nice = NULL,
  step = NULL,
  steps = NULL,
  .object = NULL
)

vl_bin_fill(
  spec,
  anchor = NULL,
  base = NULL,
  binned = NULL,
  divide = NULL,
  extent = NULL,
  maxbins = NULL,
  minstep = NULL,
  nice = NULL,
  step = NULL,
  steps = NULL,
  .object = NULL
)

vl_bin_fillOpacity(
  spec,
  anchor = NULL,
  base = NULL,
  binned = NULL,
  divide = NULL,
  extent = NULL,
  maxbins = NULL,
  minstep = NULL,
  nice = NULL,
  step = NULL,
  steps = NULL,
  .object = NULL
)

vl_bin_href(
  spec,
  anchor = NULL,
  base = NULL,
  binned = NULL,
  divide = NULL,
  extent = NULL,
  maxbins = NULL,
  minstep = NULL,
  nice = NULL,
  step = NULL,
  steps = NULL,
  .object = NULL
)

vl_bin_key(
  spec,
  anchor = NULL,
  base = NULL,
  binned = NULL,
  divide = NULL,
  extent = NULL,
  maxbins = NULL,
  minstep = NULL,
  nice = NULL,
  step = NULL,
  steps = NULL,
  .object = NULL
)

vl_bin_latitude(
  spec,
  anchor = NULL,
  base = NULL,
  binned = NULL,
  divide = NULL,
  extent = NULL,
  maxbins = NULL,
  minstep = NULL,
  nice = NULL,
  step = NULL,
  steps = NULL,
  .object = NULL
)

vl_bin_latitude2(
  spec,
  anchor = NULL,
  base = NULL,
  binned = NULL,
  divide = NULL,
  extent = NULL,
  maxbins = NULL,
  minstep = NULL,
  nice = NULL,
  step = NULL,
  steps = NULL,
  .object = NULL
)

vl_bin_longitude(
  spec,
  anchor = NULL,
  base = NULL,
  binned = NULL,
  divide = NULL,
  extent = NULL,
  maxbins = NULL,
  minstep = NULL,
  nice = NULL,
  step = NULL,
  steps = NULL,
  .object = NULL
)

vl_bin_longitude2(
  spec,
  anchor = NULL,
  base = NULL,
  binned = NULL,
  divide = NULL,
  extent = NULL,
  maxbins = NULL,
  minstep = NULL,
  nice = NULL,
  step = NULL,
  steps = NULL,
  .object = NULL
)

vl_bin_opacity(
  spec,
  anchor = NULL,
  base = NULL,
  binned = NULL,
  divide = NULL,
  extent = NULL,
  maxbins = NULL,
  minstep = NULL,
  nice = NULL,
  step = NULL,
  steps = NULL,
  .object = NULL
)

vl_bin_order(
  spec,
  anchor = NULL,
  base = NULL,
  binned = NULL,
  divide = NULL,
  extent = NULL,
  maxbins = NULL,
  minstep = NULL,
  nice = NULL,
  step = NULL,
  steps = NULL,
  .object = NULL
)

vl_bin_radius(
  spec,
  anchor = NULL,
  base = NULL,
  binned = NULL,
  divide = NULL,
  extent = NULL,
  maxbins = NULL,
  minstep = NULL,
  nice = NULL,
  step = NULL,
  steps = NULL,
  .object = NULL
)

vl_bin_radius2(
  spec,
  anchor = NULL,
  base = NULL,
  binned = NULL,
  divide = NULL,
  extent = NULL,
  maxbins = NULL,
  minstep = NULL,
  nice = NULL,
  step = NULL,
  steps = NULL,
  .object = NULL
)

vl_bin_shape(
  spec,
  anchor = NULL,
  base = NULL,
  binned = NULL,
  divide = NULL,
  extent = NULL,
  maxbins = NULL,
  minstep = NULL,
  nice = NULL,
  step = NULL,
  steps = NULL,
  .object = NULL
)

vl_bin_size(
  spec,
  anchor = NULL,
  base = NULL,
  binned = NULL,
  divide = NULL,
  extent = NULL,
  maxbins = NULL,
  minstep = NULL,
  nice = NULL,
  step = NULL,
  steps = NULL,
  .object = NULL
)

vl_bin_stroke(
  spec,
  anchor = NULL,
  base = NULL,
  binned = NULL,
  divide = NULL,
  extent = NULL,
  maxbins = NULL,
  minstep = NULL,
  nice = NULL,
  step = NULL,
  steps = NULL,
  .object = NULL
)

vl_bin_strokeDash(
  spec,
  anchor = NULL,
  base = NULL,
  binned = NULL,
  divide = NULL,
  extent = NULL,
  maxbins = NULL,
  minstep = NULL,
  nice = NULL,
  step = NULL,
  steps = NULL,
  .object = NULL
)

vl_bin_strokeOpacity(
  spec,
  anchor = NULL,
  base = NULL,
  binned = NULL,
  divide = NULL,
  extent = NULL,
  maxbins = NULL,
  minstep = NULL,
  nice = NULL,
  step = NULL,
  steps = NULL,
  .object = NULL
)

vl_bin_strokeWidth(
  spec,
  anchor = NULL,
  base = NULL,
  binned = NULL,
  divide = NULL,
  extent = NULL,
  maxbins = NULL,
  minstep = NULL,
  nice = NULL,
  step = NULL,
  steps = NULL,
  .object = NULL
)

vl_bin_text(
  spec,
  anchor = NULL,
  base = NULL,
  binned = NULL,
  divide = NULL,
  extent = NULL,
  maxbins = NULL,
  minstep = NULL,
  nice = NULL,
  step = NULL,
  steps = NULL,
  .object = NULL
)

vl_bin_theta(
  spec,
  anchor = NULL,
  base = NULL,
  binned = NULL,
  divide = NULL,
  extent = NULL,
  maxbins = NULL,
  minstep = NULL,
  nice = NULL,
  step = NULL,
  steps = NULL,
  .object = NULL
)

vl_bin_theta2(
  spec,
  anchor = NULL,
  base = NULL,
  binned = NULL,
  divide = NULL,
  extent = NULL,
  maxbins = NULL,
  minstep = NULL,
  nice = NULL,
  step = NULL,
  steps = NULL,
  .object = NULL
)

vl_bin_tooltip(
  spec,
  anchor = NULL,
  base = NULL,
  binned = NULL,
  divide = NULL,
  extent = NULL,
  maxbins = NULL,
  minstep = NULL,
  nice = NULL,
  step = NULL,
  steps = NULL,
  .object = NULL
)

vl_bin_url(
  spec,
  anchor = NULL,
  base = NULL,
  binned = NULL,
  divide = NULL,
  extent = NULL,
  maxbins = NULL,
  minstep = NULL,
  nice = NULL,
  step = NULL,
  steps = NULL,
  .object = NULL
)

vl_bin_x(
  spec,
  anchor = NULL,
  base = NULL,
  binned = NULL,
  divide = NULL,
  extent = NULL,
  maxbins = NULL,
  minstep = NULL,
  nice = NULL,
  step = NULL,
  steps = NULL,
  .object = NULL
)

vl_bin_x2(
  spec,
  anchor = NULL,
  base = NULL,
  binned = NULL,
  divide = NULL,
  extent = NULL,
  maxbins = NULL,
  minstep = NULL,
  nice = NULL,
  step = NULL,
  steps = NULL,
  .object = NULL
)

vl_bin_xError(
  spec,
  anchor = NULL,
  base = NULL,
  binned = NULL,
  divide = NULL,
  extent = NULL,
  maxbins = NULL,
  minstep = NULL,
  nice = NULL,
  step = NULL,
  steps = NULL,
  .object = NULL
)

vl_bin_xError2(
  spec,
  anchor = NULL,
  base = NULL,
  binned = NULL,
  divide = NULL,
  extent = NULL,
  maxbins = NULL,
  minstep = NULL,
  nice = NULL,
  step = NULL,
  steps = NULL,
  .object = NULL
)

vl_bin_xOffset(
  spec,
  anchor = NULL,
  base = NULL,
  binned = NULL,
  divide = NULL,
  extent = NULL,
  maxbins = NULL,
  minstep = NULL,
  nice = NULL,
  step = NULL,
  steps = NULL,
  .object = NULL
)

vl_bin_y(
  spec,
  anchor = NULL,
  base = NULL,
  binned = NULL,
  divide = NULL,
  extent = NULL,
  maxbins = NULL,
  minstep = NULL,
  nice = NULL,
  step = NULL,
  steps = NULL,
  .object = NULL
)

vl_bin_y2(
  spec,
  anchor = NULL,
  base = NULL,
  binned = NULL,
  divide = NULL,
  extent = NULL,
  maxbins = NULL,
  minstep = NULL,
  nice = NULL,
  step = NULL,
  steps = NULL,
  .object = NULL
)

vl_bin_yError(
  spec,
  anchor = NULL,
  base = NULL,
  binned = NULL,
  divide = NULL,
  extent = NULL,
  maxbins = NULL,
  minstep = NULL,
  nice = NULL,
  step = NULL,
  steps = NULL,
  .object = NULL
)

vl_bin_yError2(
  spec,
  anchor = NULL,
  base = NULL,
  binned = NULL,
  divide = NULL,
  extent = NULL,
  maxbins = NULL,
  minstep = NULL,
  nice = NULL,
  step = NULL,
  steps = NULL,
  .object = NULL
)

vl_bin_yOffset(
  spec,
  anchor = NULL,
  base = NULL,
  binned = NULL,
  divide = NULL,
  extent = NULL,
  maxbins = NULL,
  minstep = NULL,
  nice = NULL,
  step = NULL,
  steps = NULL,
  .object = NULL
)

vl_bin_facet(
  spec,
  anchor = NULL,
  base = NULL,
  binned = NULL,
  divide = NULL,
  extent = NULL,
  maxbins = NULL,
  minstep = NULL,
  nice = NULL,
  step = NULL,
  steps = NULL,
  .object = NULL
)

vl_bin_row(
  spec,
  anchor = NULL,
  base = NULL,
  binned = NULL,
  divide = NULL,
  extent = NULL,
  maxbins = NULL,
  minstep = NULL,
  nice = NULL,
  step = NULL,
  steps = NULL,
  .object = NULL
)

vl_bin_column(
  spec,
  anchor = NULL,
  base = NULL,
  binned = NULL,
  divide = NULL,
  extent = NULL,
  maxbins = NULL,
  minstep = NULL,
  nice = NULL,
  step = NULL,
  steps = NULL,
  .object = NULL
)

Arguments

spec

An input vega-lite spec

anchor

A value in the binned domain at which to anchor the bins, shifting the bin boundaries if necessary to ensure that a boundary aligns with the anchor value.

Default value: the minimum bin extent value

base

The number base to use for automatic bin determination (default is base 10).

Default value: 10

binned

When set to true, Vega-Lite treats the input data as already binned.

divide

Scale factors indicating allowable subdivisions. The default value is [5, 2], which indicates that for base 10 numbers (the default base), the method may consider dividing bin sizes by 5 and/or 2. For example, for an initial step size of 10, the method can check if bin sizes of 2 (= 10/5), 5 (= 10/2), or 1 (= 10/(5*2)) might also satisfy the given constraints.

Default value: [5, 2]

extent

A two-element ([min, max]) array indicating the range of desired bin values.

maxbins

Maximum number of bins.

Default value: 6 for row, column and shape channels; 10 for other channels

minstep

A minimum allowable step size (particularly useful for integer values).

nice

If true, attempts to make the bin boundaries use human-friendly boundaries, such as multiples of ten.

Default value: true

step

An exact step size to use between bins.

Note: If provided, options such as maxbins will be ignored.

steps

An array of allowable step sizes to choose from.

.object

Directly input an object, rather than creating one via the other arguments. Should not be used in conjunction with the other arguments other than spec. Objects can be of type: BinParams

Value

A modified Vega-Lite Spec